requirements.txt
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
click==8.2.0
cryptography==45.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.10
iniconfig==2.1.0
init==0.1.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mock==3.14.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rfc3986==1.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
terminal==0.4.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
